Move AwaitableEvent to core & refactor
