+++ src/index.ts
@@ -1,8 +1,9 @@
 export { default as action } from './action';
 export { default as asyncState, type AsyncState } from './async-state';
+export { AwaitableEvent } from './core';
 export { default as effect } from './effect';
 export { default as familyState } from './family-state';
-export { AwaitableEvent, delay, fork, getAggregatedAsyncStatus, rejectAfter } from './lib';
+export { delay, fork, getAggregatedAsyncStatus, rejectAfter } from './lib';
 export { scenario, scenarioOnce, scenarioOnEvery } from './scenario';
 export { default as selector } from './selector';
 export { default as state, type State } from './state';
