+++ src/lib/index.ts
@@ -1,4 +1,3 @@
-export { default as AwaitableEvent } from './AwaitableEvent';
 export { default as delay } from './delay';
 export { default as flush } from './flush';
 export { default as fork } from './fork';
